+++ src/day1_2/Position.java
@@ -2,49 +2,49 @@ package day1_2;
 
 // Step.2-3 クラス
 public class Position {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 
 	public Position() {
 		this(0, 0);
 	}
 
-	public Position(int x, int y) {
+	public Position(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
-	public void set(int x, int y) {
+	public void set(double x, double y) {
 		this.setX(x);
 		this.setY(y);
 	}
 
-	public void addX(int vx) {
+	public void addX(double vx) {
 		this.x += vx;
 	}
 
-	public void addY(int vy) {
+	public void addY(double vy) {
 		this.y += vy;
 	}
 
-	public void add(int vx, int vy) {
+	public void add(double vx, double vy) {
 		this.addX(vx);
 		this.addY(vy);
 	}
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 }
