+++ src/day1_3/Rect.java
@@ -1,10 +1,10 @@
 package day1_3;
 
 public class Rect extends Figure {
-	private int width;
-	private int height;
+	private double width;
+	private double height;
 
-	public Rect(int x, int y, int width, int height) {
+	public Rect(double x, double y, double width, double height) {
 		super(x, y);
 		this.width = width;
 		this.height = height;
